add dummy node
